% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Title birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{total number of people in the room}
}
\value{
probability of having 2 or more people in the with the same birthday
}
\description{
Title birthday
}
\examples{
birthday(20)
}
