% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSS.R
\name{MSS}
\alias{MSS}
\title{Model Sum of Squares}
\usage{
MSS(yhat, mean)
}
\arguments{
\item{yhat}{Estimate for y}

\item{mean}{Mean of y}
}
\value{
The Model Sum of Squares given a vector estimate for y
}
\description{
Model Sum of Squares
}
