% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of trials}

\item{x}{a vector of a sample}

\item{fun}{the statistic of interest}

\item{alpha}{confidence level}

\item{cx}{hex}

\item{...}{}
}
\value{
a barplot of the density for the desired statistic given a sample. calculates point and interval estimates
}
\description{
Title myboot2
}
\examples{
myboot2(iter=10000,x=rnorm(30),fun="mean",alpha=0.05)
}
