% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Title mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of observations per sample}

\item{iter}{number of samples}

\item{p}{probability of success in each trial}

\item{...}{}
}
\value{
histogram of the mean of the samples
}
\description{
Title mycltb
}
\examples{
mycltb(n=10,iter=1000,p=0.5)
}
