% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My first function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of y and x.
}
\description{
My first function
}
\examples{
myfirstfunction(1:10)
}
