% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{My multinomial}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of iterations to conduct}

\item{n}{number of items to sample}

\item{p}{probability of each outcome}
}
\value{
barplot of relative frequency for a multinomial distribution
}
\description{
My multinomial
}
\examples{
mymult(iter=100,n=10,p=c(1,1,1,1)/4)
}
