% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean for the normal distribution}

\item{sigma}{the standard deviation for the normal distribution}

\item{a}{the x value to calculate the percentile}
}
\value{
a normal distribution with the P(X<a)
}
\description{
myncurve
}
\examples{
myncurve(mu=0,sigma=1,a=0)
}
