% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title number of tickets}
\usage{
ntickets(N = 200, gamma = 0.8, p = 0.95)
}
\arguments{
\item{N}{number of seats on the plane}

\item{gamma}{acceptable probability of loss}

\item{p}{probability of a seat being filled}
}
\value{
number of tickets to sell
}
\description{
Title number of tickets
}
\examples{
ntickets(N=400,gamma=0.02,p=0.95)
}
